[Fix] Fix possible (but very unlikely) infinite loop in A*
If the record holder was ever determined to be null,
we would simply continue to the next iteration of the loop, without
changing any data.
This would lead to the most promising node still being null.
GetMostPromisingNode will never return null if the list is not empty,
which is the condition for the loop, but better safe than sorry.